Refactor settings views and remove placeholders
