--- CouplesCount/Services.swift
@@ -1,7 +1,0 @@
-//
-//  Services.swift
-//  CouplesCount
-//
-//  Created by Michael Bubienczyk on 2025-08-18.
-//
-
