--- CouplesCount/ViewModels.swift
@@ -1,7 +1,0 @@
-//
-//  ViewModels.swift
-//  CouplesCount
-//
-//  Created by Michael Bubienczyk on 2025-08-18.
-//
-
